requirements.txt
streamlit
langchain==0.0.153
openai
python-dotenv
deeplake
tiktoken
faiss-cpu
